requirements.txt
langchain
langchain-google-genai
langchain-community
pydantic
fpdf
jinja2
pypdf
sqlalchemy
python-dotenv
fastapi
uvicorn
python-multipart
thefuzz
azure-ai-documentintelligence
azure-core
reportlab
